requirements.txt
python-benedict~=0.33.0
requests~=2.32.0
numpy
python-dotenv~=1.0.0